+++ src/interfaces/ITintolmarketServerSkel.java
@@ -11,7 +11,7 @@ public interface ITintolmarketServerSkel {
 	
 	public String viewWallet(String userID);
 	
-	public String classifyWine(String wine, int stars);
+	public String classifyWine(String wine, float stars);
 
 	public String sendMessage(String user, String userFrom, String message) ;
